+++ tests/TestDataService/TestDataService.spec.ts
@@ -1,4 +1,3 @@
-/* eslint-disable playwright/no-conditional-expect */
 import {
     test,
     expect,